requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0
pydantic>=2.6,<3.0
mistralai>=1.0,<2.0
langchain>=0.3
langchain-core>=0.3
langchain-text-splitters>=0.3
langchain-mistralai>=0.2
faiss-cpu>=1.7,<2.0
numpy>=1.26,<3.0
requests>=2.32,<3.0
python-dotenv>=1.0,<2.0
ragas>=0.2,<0.3
datasets>=2.18
pytest>=8.0
httpx>=0.27
